Fix typos

+++ src/api/java/io/github/cvc5/Solver.java
@@ -2101,7 +2101,7 @@ public class Solver implements IPointer
     return proofToString(pointer, proofPointers, format.getValue());
   }
 
-  private native String proofToString(long pointer, long[] proofs, long format);
+  private native String proofToString(long pointer, long[] proofs, int format);
 
   /**
    * Prints a vector of proofs into a string with a slected proof format mode.
@@ -2124,7 +2124,7 @@ public class Solver implements IPointer
     return proofToString(pointer, proofPointers, format.getValue(), component.getValue());
   }
 
-  private native String proofToString(long pointer, long[] proofs, long format, long component);
+  private native String proofToString(long pointer, long[] proofs, int format, int component);
 
   /**
    * Get the value of the given term in the current model.
